Refactoring/simplification part 1
