--- vertx-mail-client/src/test/java/io/vertx/ext/mail/JavascriptTest.java
@@ -1,19 +1,0 @@
-package io.vertx.ext.mail;
-
-import io.vertx.test.lang.js.JSRunner;
-
-import org.junit.Test;
-
-/**
- * @author <a href="http://oss.lehmann.cx/">Alexander Lehmann</a>
- *
- * run a javascript test file in e.g. eclipse
- */
-public class JavascriptTest {
-
-  @Test
-  public void runJavascriptTest() throws Exception {
-    new JSRunner().run("javascript/mailtest.js", "*");
-  }
-
-}
